tunelator/requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.21.33
botocore==1.24.33
CacheControl==0.12.10
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
Deprecated==1.2.13
Django==3.2
django-admin-sortable==2.2.4
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-mirage-field==1.4.0
django-polymorphic==3.1.0
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
inotify==0.2.10
install==1.3.5
jmespath==1.0.0
kombu==5.2.4
Markdown==3.3.6
msgpack==1.0.3
nose==1.3.7
packaging==21.3
prompt-toolkit==3.0.28
proto-plus==1.20.3
protobuf==3.20.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.2.1
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.7.0
